# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Royvento event management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Invalid request
        '500':
          description: Server error
...
  /users/me:
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
  /users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
...
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    Role:
      type: string
      enum: [user, vendor, admin]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        phone:
          type: string
        about:
          type: string
        profileImage:
          type: string
        referralCode:
          type: string
        referredBy:
          type: integer
          nullable: true
        points:
          type: integer
        createdAt:
          type: string
      required: [id, email, name, role, phone, about, profileImage, referralCode, points, createdAt]
...
    UpdateRoleBody:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/Role"
      required: [role]
    UpdateMeBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        phone:
          type: string
          maxLength: 50
        about:
          type: string
          maxLength: 2000
        profileImage:
          type: string
          maxLength: 2048
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
...